The structure is completed
