--- src/FileManager.java
@@ -1,11 +1,0 @@
-package src;
-
-public class FileManager {
-  public String readFile(String filePath) {
-    // File reading logic
-    return filePath;
-  }
-  public void writeFile(String content, String filePath) {
-    // Logic for writing a file
-  }
-}
