--- src/Main.java
@@ -1,7 +1,0 @@
-package src;
-
-public class Main {
-  public static void main(String[] args) {
-    System.out.println("123");
-  }
-}
