--- src/StatisticalAnalyzer.java
@@ -1,9 +1,0 @@
-package src;
-
-public class StatisticalAnalyzer {
-  public int findMostLikelyShift(String encryptedText,
-                                 char[] alphabet, String representativeText) {
-    // Statistical analysis logic to determine the shift
-    return 0;
-  }
-}
