+++ src/cipher/Cipher.java
@@ -1,0 +1,16 @@
+package src.cipher;
+
+public class Cipher {
+  private char[] alphabet;
+  public Cipher(char[] alphabet) {
+    this.alphabet = alphabet;
+  }
+  public String encrypt(String text, int shift) {
+    // Encryption logic
+    return null;
+  }
+  public String decrypt(String encryptedText, int shift) {
+    // Decryption logic
+    return null;
+  }
+}
