+++ src/statistic/StatisticalAnalyzer.java
@@ -1,0 +1,12 @@
+package src.statistic;
+
+public class StatisticalAnalyzer {
+  public int findMostLikelyShift(
+      String encryptedText,
+      char[] alphabet,
+      String representativeText
+  ) {
+    // Statistical analysis logic to determine the shift
+    return 0;
+  }
+}
